refactor(animations): unify WindowAnimation into compositor-common
Promote kwin-effect's local ManagedAnimation struct into compositor-common's
WindowAnimation, collapsing the duplicate so any future compositor plugin
(Wayfire, Hyprland, ...) inherits the same styled animation state without
re-deriving it from an Effect-specific superset. Follow-up from the v3
reintegration commit (b4351d06).

## compositor-common

- easingcurve.h now includes `../common/animationstyle.h` and
  `../common/springparams.h` so SpringAnimation / AnimationStyle are
  visible to the shared struct. These headers stay in src/common/ so
  daemon code that already includes them isn't disturbed; compositor-common
  uses relative-path includes rather than adding new include directories.
- New SpringAnimation struct (damped harmonic oscillator) — moved out of
  kwin-effect verbatim. Underdamped / critically damped / overdamped
  branches with the same near-zero and finite-value guards.
- WindowAnimation replaces the plain `EasingCurve easing` field with
  `std::variant<EasingCurve, SpringAnimation> timing`, and gains
  cachedSpringDuration, style, styleParam, shaderPath, vertexShaderPath,
  and shaderSubdivisions. Helpers isSpring() / usesOpacity() expose the
  state needed by paint-path dispatch. updateProgress() / isComplete()
  branch on the variant.

## kwin-effect

- Deletes the local SpringAnimation / ManagedAnimation definitions and
  their implementations (~170 lines). `src/common/*` includes go away
  too, since WindowAnimation now pulls them in via easingcurve.h.
- WindowAnimator::m_animations and every private transform method take
  `WindowAnimation` again — no behavioral change, just a rename. The
  `AnimationInfo` snapshot, `applyGeometryOnly`, per-style transforms,
  and overshoot-aware animationBounds() continue to work unchanged
  against the unified struct.
- AnimationParams stays local because `fromJson` depends on ConfigKeys
  accessors that live in the daemon module; it already represents the
  DTO boundary between daemon JSON and effect runtime state.

## animation_math / tests

- createSnapAnimation() now sets `anim.timing = config.easing` (assigns
  into the variant slot). Functional behavior is identical.
- test_compositor_common only inspects startPosition / startSize /
  targetGeometry / duration on the returned animation, so no test
  touch-up was needed. 78/78 tests pass.

Net diff: -52 lines.
