+++ kwin-effect/windowanimator.cpp
@@ -21,91 +21,7 @@ namespace PlasmaZones {
 
 Q_DECLARE_LOGGING_CATEGORY(lcEffect)
 
-// ═══════════════════════════════════════════════════════════════════════════════
-// SpringAnimation
-// ═══════════════════════════════════════════════════════════════════════════════
-
-qreal SpringAnimation::evaluate(qreal t) const
-{
-    if (t <= 0.0)
-        return 0.0;
-
-    const qreal safeStiffness = qMax(1.0, stiffness);
-    const qreal omega = qSqrt(safeStiffness); // natural frequency (mass = 1)
-    const qreal zeta = qMax(0.0, dampingRatio);
-
-    if (zeta < 1.0) {
-        // Underdamped: oscillates
-        const qreal omegaD = omega * qSqrt(1.0 - zeta * zeta);
-        // Guard near-critical damping where omegaD approaches zero — fall through
-        // to the critically-damped formula to avoid division by near-zero.
-        if (omegaD < 1e-6) {
-            const qreal decay = qExp(-omega * t);
-            return 1.0 - decay * (1.0 + (omega - initialVelocity) * t);
-        }
-        const qreal decay = qExp(-zeta * omega * t);
-        const qreal v0 = initialVelocity;
-        const qreal sinTerm = (zeta * omega - v0) / omegaD;
-        return 1.0 - decay * (qCos(omegaD * t) + sinTerm * qSin(omegaD * t));
-    } else if (qFuzzyCompare(zeta, 1.0)) {
-        // Critically damped
-        const qreal decay = qExp(-omega * t);
-        return 1.0 - decay * (1.0 + (omega - initialVelocity) * t);
-    } else {
-        // Overdamped
-        const qreal disc = qSqrt(zeta * zeta - 1.0);
-        const qreal s1 = -omega * (zeta + disc);
-        const qreal s2 = -omega * (zeta - disc);
-        if (qAbs(s2 - s1) < 1e-6) {
-            const qreal decay = qExp(-omega * t);
-            return 1.0 - decay * (1.0 + (omega - initialVelocity) * t);
-        }
-        const qreal c2 = (initialVelocity - s1) / (s2 - s1);
-        const qreal c1 = 1.0 - c2;
-        const qreal result = 1.0 - (c1 * qExp(s1 * t) + c2 * qExp(s2 * t));
-        if (!std::isfinite(result))
-            return 1.0;
-        return result;
-    }
-}
-
-bool SpringAnimation::isSettled(qreal t) const
-{
-    if (t <= 0.0)
-        return false;
-    const qreal pos = evaluate(t);
-    const qreal dt = 0.001;
-    const qreal vel = (evaluate(t + dt) - pos) / dt;
-    const qreal safeEps = qMax(0.00001, epsilon);
-    return qAbs(pos - 1.0) < safeEps && qAbs(vel) < safeEps * 10.0;
-}
-
-qreal SpringAnimation::estimatedDuration() const
-{
-    // Forward scan: require 3 consecutive settled samples to avoid
-    // false positives from underdamped springs briefly crossing epsilon.
-    constexpr qreal dt = 0.005;
-    constexpr qreal maxT = 5.0;
-    constexpr int requiredConsecutive = 3;
-    int consecutive = 0;
-    for (qreal t = dt; t <= maxT; t += dt) {
-        if (isSettled(t)) {
-            if (++consecutive >= requiredConsecutive)
-                return t - (requiredConsecutive - 1) * dt;
-        } else {
-            consecutive = 0;
-        }
-    }
-    return maxT;
-}
-
-QString SpringAnimation::toString() const
-{
-    return QStringLiteral("spring(damping=%1,stiffness=%2,epsilon=%3)")
-        .arg(dampingRatio, 0, 'f', 2)
-        .arg(stiffness, 0, 'f', 0)
-        .arg(epsilon, 0, 'g', 4);
-}
+// SpringAnimation physics live in src/compositor-common/easingcurve.cpp.
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // AnimationParams
@@ -193,7 +109,7 @@ bool WindowAnimator::startAnimation(KWin::EffectWindow* window, const QPointF& o
         --m_opacityAnimationCount;
     }
 
-    ManagedAnimation anim;
+    WindowAnimation anim;
     anim.startPosition = oldPosition;
     anim.startSize = oldSize;
     anim.targetGeometry = targetGeometry;
@@ -400,7 +316,7 @@ std::optional<WindowAnimator::AnimationInfo> WindowAnimator::animationInfo(KWin:
                          it->shaderPath,     it->vertexShaderPath, it->shaderSubdivisions};
 }
 
-void WindowAnimator::applyGeometryInterpolation(KWin::EffectWindow* window, const ManagedAnimation& anim,
+void WindowAnimator::applyGeometryInterpolation(KWin::EffectWindow* window, const WindowAnimation& anim,
                                                 KWin::WindowPaintData& data, qreal slideFraction) const
 {
     // Translate: desired visual position minus actual (frameGeometry is already at target).
@@ -423,13 +339,13 @@ void WindowAnimator::applyGeometryInterpolation(KWin::EffectWindow* window, cons
     }
 }
 
-void WindowAnimator::applyMorphTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+void WindowAnimator::applyMorphTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                                          KWin::WindowPaintData& data) const
 {
     applyGeometryInterpolation(window, anim, data);
 }
 
-void WindowAnimator::applySlideTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+void WindowAnimator::applySlideTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                                          KWin::WindowPaintData& data) const
 {
     applyGeometryInterpolation(window, anim, data);
@@ -437,7 +353,7 @@ void WindowAnimator::applySlideTransform(KWin::EffectWindow* window, const Manag
     data.multiplyOpacity(qBound(0.0, anim.cachedProgress / 0.3, 1.0));
 }
 
-void WindowAnimator::applyPopinTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+void WindowAnimator::applyPopinTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                                          KWin::WindowPaintData& data) const
 {
     const qreal p = anim.cachedProgress;
@@ -459,7 +375,7 @@ void WindowAnimator::applyPopinTransform(KWin::EffectWindow* window, const Manag
     data.multiplyOpacity(qBound(0.0, p, 1.0));
 }
 
-void WindowAnimator::applySlideFadeTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+void WindowAnimator::applySlideFadeTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                                              KWin::WindowPaintData& data) const
 {
     const qreal slideFraction = qBound(0.0, anim.styleParam, 1.0);
