+++ kwin-effect/windowanimator.h
@@ -3,9 +3,6 @@
 
 #pragma once
 
-#include "../src/common/animationstyle.h"
-#include "../src/common/springparams.h"
-
 #include <easingcurve.h>
 
 #include <QHash>
@@ -18,7 +15,6 @@
 #include <QString>
 #include <chrono>
 #include <optional>
-#include <variant>
 
 namespace KWin {
 class EffectWindow;
@@ -27,33 +23,18 @@ class WindowPaintData;
 
 namespace PlasmaZones {
 
-// EasingCurve and AnimationConfig live in src/compositor-common/easingcurve.h.
-// This effect adds SpringAnimation (physics-based) plus style-aware window
-// transitions on top of that shared baseline.
-
-/**
- * @brief Damped harmonic oscillator animation (niri-inspired spring physics)
- *
- * Inherits the persistent config fields (dampingRatio, stiffness, epsilon) from
- * SpringParams and adds runtime-only fields (initialVelocity) plus physics
- * evaluation methods. Unlike EasingCurve, spring animations have no fixed
- * duration — they converge based on physics parameters.
- */
-struct SpringAnimation : SpringParams
-{
-    qreal initialVelocity = 0.0; ///< Initial velocity (e.g. from gesture release)
-
-    qreal evaluate(qreal t) const;
-    bool isSettled(qreal t) const;
-    qreal estimatedDuration() const;
-    QString toString() const;
-};
+// EasingCurve, SpringAnimation, WindowAnimation, and AnimationConfig live in
+// src/compositor-common/easingcurve.h (shared across compositor plugins).
+// This effect provides the runtime animator that drives them and the
+// AnimationParams DTO for per-event profile resolution.
 
 /**
  * @brief Resolved per-event animation parameters
  *
- * Produced by resolving a profile from AnimationProfileTree (see
- * src/core/animationprofile.h) for a specific event like "snapIn" or "snapOut".
+ * Produced by parsing a resolved AnimationProfile JSON returned from the
+ * daemon (Settings.resolveAnimationProfile D-Bus method). Held locally in
+ * kwin-effect because parsing depends on ConfigKeys accessors that live
+ * in the daemon module.
  */
 struct AnimationParams
 {
@@ -71,114 +52,6 @@ struct AnimationParams
     /// Parse from a resolved AnimationProfile JSON object (schema from
     /// AnimationProfile::toJson() on the daemon side).
     static AnimationParams fromJson(const QJsonObject& obj);
-};
-
-/**
- * @brief Runtime state of a single active window animation
- *
- * Superset of compositor-common's WindowAnimation, extended with spring
- * physics, style, and shader metadata. Kept local to kwin-effect so
- * compositor-common stays free of per-effect extensions.
- */
-struct ManagedAnimation
-{
-    QPointF startPosition;
-    QSizeF startSize;
-    QRect targetGeometry;
-    std::chrono::milliseconds startTime{-1};
-    qreal duration = 150.0; ///< ms (ignored for spring)
-    std::variant<EasingCurve, SpringAnimation> timing;
-    qreal cachedProgress = 0.0;
-    qreal cachedSpringDuration = -1.0; ///< Precomputed estimatedDuration() for springs
-
-    AnimationStyle style = AnimationStyle::Morph;
-    qreal styleParam = 0.87;
-    QString shaderPath;
-    QString vertexShaderPath;
-    int shaderSubdivisions = 1;
-
-    bool isValid() const
-    {
-        return startTime.count() >= 0;
-    }
-
-    bool isSpring() const
-    {
-        return std::holds_alternative<SpringAnimation>(timing);
-    }
-
-    /// Style uses opacity → paint path must enable translucent rendering
-    bool usesOpacity() const
-    {
-        return style == AnimationStyle::Slide || style == AnimationStyle::Popin || style == AnimationStyle::SlideFade
-            || style == AnimationStyle::FadeIn || style == AnimationStyle::SlideUp || style == AnimationStyle::ScaleIn;
-    }
-
-    void updateProgress(std::chrono::milliseconds presentTime)
-    {
-        if (startTime.count() < 0) {
-            startTime = presentTime;
-            cachedProgress = 0.0;
-            return;
-        }
-        const qreal elapsedMs = qreal((presentTime - startTime).count());
-
-        if (isSpring()) {
-            const auto& spring = std::get<SpringAnimation>(timing);
-            cachedProgress = spring.evaluate(elapsedMs / 1000.0);
-        } else {
-            if (duration <= 0.0) {
-                cachedProgress = 1.0;
-                return;
-            }
-            const qreal t = qMin(1.0, elapsedMs / duration);
-            cachedProgress = std::get<EasingCurve>(timing).evaluate(t);
-        }
-    }
-
-    qreal progress() const
-    {
-        return cachedProgress;
-    }
-
-    bool isComplete(std::chrono::milliseconds presentTime) const
-    {
-        if (startTime.count() < 0) {
-            return false;
-        }
-        const qreal elapsedMs = qreal((presentTime - startTime).count());
-        if (isSpring()) {
-            return std::get<SpringAnimation>(timing).isSettled(elapsedMs / 1000.0);
-        }
-        if (duration <= 0.0) {
-            return true;
-        }
-        return elapsedMs >= duration;
-    }
-
-    QPointF currentVisualPosition() const
-    {
-        const qreal p = cachedProgress;
-        const qreal tx = targetGeometry.x();
-        const qreal ty = targetGeometry.y();
-        return QPointF(startPosition.x() + (tx - startPosition.x()) * p,
-                       startPosition.y() + (ty - startPosition.y()) * p);
-    }
-
-    QSizeF currentVisualSize() const
-    {
-        const qreal p = cachedProgress;
-        const qreal tw = targetGeometry.width();
-        const qreal th = targetGeometry.height();
-        return QSizeF(startSize.width() + (tw - startSize.width()) * p,
-                      startSize.height() + (th - startSize.height()) * p);
-    }
-
-    bool hasScaleChange() const
-    {
-        return qAbs(startSize.width() - targetGeometry.width()) > 1.0
-            || qAbs(startSize.height() - targetGeometry.height()) > 1.0;
-    }
 };
 
 /**
@@ -319,19 +192,19 @@ Q_SIGNALS:
 
 private:
     // Shared translate + scale interpolation used by Morph, Slide, SlideFade
-    void applyGeometryInterpolation(KWin::EffectWindow* window, const ManagedAnimation& anim,
+    void applyGeometryInterpolation(KWin::EffectWindow* window, const WindowAnimation& anim,
                                     KWin::WindowPaintData& data, qreal slideFraction = 1.0) const;
 
-    void applyMorphTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+    void applyMorphTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                              KWin::WindowPaintData& data) const;
-    void applySlideTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+    void applySlideTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                              KWin::WindowPaintData& data) const;
-    void applyPopinTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+    void applyPopinTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                              KWin::WindowPaintData& data) const;
-    void applySlideFadeTransform(KWin::EffectWindow* window, const ManagedAnimation& anim,
+    void applySlideFadeTransform(KWin::EffectWindow* window, const WindowAnimation& anim,
                                  KWin::WindowPaintData& data) const;
 
-    QHash<KWin::EffectWindow*, ManagedAnimation> m_animations;
+    QHash<KWin::EffectWindow*, WindowAnimation> m_animations;
     int m_opacityAnimationCount = 0;
     bool m_enabled = true;
     qreal m_duration = 150.0;
